pyproject.toml
[tool.poetry]
name = "pp-hw-01"
version = "0.1.0"
description = "otus python professional homework_01"
authors = ["valed-dm <dmvaled@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "deco" },
    { include = "log_analyzer" }
]

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
bs4 = "^0.0.1"
lxml = "^4.9.2"
pylint = "^2.17.4"
pytest = "^7.3.1"
selenium = "^4.10.0"
snapshottest = "^0.6.0"
webdriver-manager = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
